tests/handlers/SshServerHandler.test.ts
// Import necessary modules and mock setups
import { Client } from 'ssh2';
import SshServerHandler from '../../src/handlers/SshServerHandler';

// Assuming the mock setup is similar to what you've already provided
// Mock setup for SFTP and other mocked functionalities

jest.mock('ssh2', () => ({
  // Your existing mock setup
}));

// Utility functions for mock setups (if any)

describe('SSH Server Handler Tests', () => {
  // Use the global variable to determine if tests should be run or skipped
  const runTests = global.shouldRunSshTests;

  // Conditional test execution based on the global variable
  (runTests ? describe : describe.skip)('SSH Server Handler', () => {
    beforeAll(() => {
      // Setup before running tests, if necessary
    });

    afterAll(() => {
      // Cleanup after tests, if necessary
    });

    it('executes a command successfully', async () => {
      // Your test logic for a successful command execution
    });

    it('handles command execution error and returns stderr', async () => {
      // Your test logic for handling command execution error
      expect.assertions(1); // Ensure that one assertion is called
      const sshServerHandler = new SshServerHandler({host: 'example.com', username: 'user'});
      // Mock setups for failure scenario

      try {
        await sshServerHandler.executeCommand('failing command');
      } catch (error) {
        const typedError = error as Error; // Type assertion for error
        expect(typedError.message).toMatch(/Command execution failed/); // Assertion
      }
    });

    // More tests as needed...
  });
});
